pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "dutchbay-finmodel"
version = "1.0.0"
description = "Dutch Bay 150 MW Wind Farm - hardened, lender-grade financial model."
readme = "README.md"
requires-python = ">=3.9"
license = { text = "Proprietary" }

dependencies = [
    "numpy>=1.24.0",
    "pandas>=2.0.0",
    "numpy-financial>=1.0.0",
    "pyyaml>=6.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-cov>=4.1.0",
    "hypothesis>=6.82.0",
    "black>=24.0.0",
    "ruff>=0.4.0",
    "mypy>=1.8.0",
    "isort>=5.13.0",
    "pre-commit>=3.5.0",
]

[tool.setuptools.packages.find]
where = ["."]
include = ["dutchbay_finmodel*"]
